fix: 계좌 값 가져와 보유 주식 표출

+++ src/main/java/org/cjlee/auto/autojoosik/autojoosik/stockDashboard/bean/VirtualAssetVO.java
@@ -30,5 +30,5 @@ public class VirtualAssetVO {
 
     private String stkNm;
 
-    private Long curPrc;
+    private Integer curPrc;
 }
